tests/wpt/tests/css/css-gaps/animation/gap-decorations-width-neutral-keyframe-005.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>gap decorations column-rule-width neutral keyframe</title>
    <link rel="help" href="https://drafts.csswg.org/css-gaps-1/#column-column-rule-width">
    <meta name="assert" content="gap decorations column-rule-width supports neutral keyframe.">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/css/support/interpolation-testcommon.js"></script>
  </head>
  <body>
    <div id="target"></div>
    <script>
      test(() => {
        target.style.columnRuleStyle = 'solid';
        target.style.columnRuleWidth= '10px 15px 20px 25px';
        var animation = target.animate([{}, {columnRuleWidth: '20px 25px 30px 35px'}], 1000);
        animation.pause();
        animation.currentTime = 500;
        assert_equals(getComputedStyle(target).columnRuleWidth, '15px 20px 25px 30px');
      }, 'gap decorations column-rule-width supports neutral keyframe.');
    </script>
  </body>
</html>
